Refactor sharing code out of identity page
